+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 ### 02/02/2021
+-   Add simple websocket server.
+-   Add `Connection` struct.
 -   Change find user routes to query url.
 -   Add users endpoints and structs.
 -   Add database connection with SQLx and PostgreSQL.